working login however not pushing data to database, removed passport js
