+++ controllers/api/index.js
@@ -2,7 +2,7 @@ const router = require('express').Router();
 const userRoutes = require('./userRoutes');
 // const projectRoutes = require('./projectRoutes');
 
-router.use('/users', userRoutes);
+router.use('/user', userRoutes);
 // router.use('/projects', projectRoutes);
 
 module.exports = router;
