+++ package.json
@@ -23,15 +23,11 @@
     "dotenv": "^10.0.0",
     "eslint": "^8.2.0",
     "express": "^4.17.1",
-    "express-flash": "0.0.2",
     "express-handlebars": "^5.3.4",
     "express-session": "^1.17.2",
     "handlebars": "^4.7.7",
-    "method-override": "^3.0.0",
     "mysql2": "^2.2.5",
     "nodemon": "^2.0.14",
-    "passport": "^0.5.0",
-    "passport-local": "^1.0.0",
     "prettier": "^2.4.1",
     "sequelize": "^6.3.5"
   }
