+++ views/myaccount.handlebars
@@ -1,24 +1,16 @@
-<nav>
-    <h1>My Account</h1>
-    <ul>
-        <a href ="/"><li>Homepage</li></a>
-        <li>Search a City</li>
-    </ul>
-</nav>
-
 <div>
-    <div>
-        <p1>User Information Here</p1>
-    </div>
+  <div>
+    <p1>User Information Here</p1>
+</div>
 </div>
 <div>
-    <div>
-        <p1>Add a review</p1>
-    </div>
+  <div>
+    <p1>Add a review</p1>
+  </div>
 </div>
 
 <div>
-    <div>
-        <p1>My Reviews</p1>
-    </div>
+  <div>
+    <p1>My Reviews</p1>
+  </div>
 </div>
