+++ views/register.handlebars
@@ -81,7 +81,7 @@
   <div class="container signin text-center">
     <p>
       Already have an account?
-      <a href="homepage#login">
+      <a href="/homepage#login">
         Sign in
       </a>
       .
